Move model transform tests into `core`
